refactor seed menu and migrations
