+++ menu_category/seed/menu_category_data.py
@@ -1,0 +1,35 @@
+parent_category_data = {
+    'cat001': {'title': 'Food', 'description': 'Main food menu'},
+    'cat002': {'title': 'Drinks', 'description': 'Beverage options'},
+    'cat003': {'title': 'Desserts', 'description': 'Sweet dishes'},
+    'cat004': {'title': 'Appetizers', 'description': 'Starters and light snacks'},
+    'cat005': {'title': 'Specials', 'description': 'Seasonal or chef\'s specials'},
+}
+
+child_category_data = {
+    'cat001': [
+        {'title': 'Rice Dishes', 'description': 'Meals served with rice'},
+        {'title': 'Noodle Dishes', 'description': 'Meals with various types of noodles'},
+        {'title': 'Grilled Items', 'description': 'Grilled meats and vegetables'},
+    ],
+    'cat002': [
+        {'title': 'Hot Drinks', 'description': 'Coffee, tea, and more'},
+        {'title': 'Cold Drinks', 'description': 'Iced beverages and soft drinks'},
+        {'title': 'Smoothies', 'description': 'Fruit and yogurt-based drinks'},
+    ],
+    'cat003': [
+        {'title': 'Cakes', 'description': 'Various cake options'},
+        {'title': 'Ice Cream', 'description': 'Classic and custom flavors'},
+        {'title': 'Pastries', 'description': 'Tarts, croissants, and more'},
+    ],
+    'cat004': [
+        {'title': 'Finger Food', 'description': 'Bite-sized snacks'},
+        {'title': 'Soups', 'description': 'Warm and hearty soups'},
+        {'title': 'Salads', 'description': 'Fresh and healthy salads'},
+    ],
+    'cat005': [
+        {'title': 'Weekly Specials', 'description': 'Updated every week'},
+        {'title': 'Holiday Specials', 'description': 'Festive themed dishes'},
+        {'title': 'Limited Time Offers', 'description': 'Time-sensitive dishes'},
+    ],
+}
